serene_proxy/requirements.txt
serene_metadata
Twisted==15.5.0
python-ldap>=2.4.25
pykerberos
